Refactors to avoid race conditions when quitting
