+++ src/components/QuickExitButton.tsx
@@ -4,8 +4,8 @@ import * as FlexWebChat from '@twilio/flex-webchat-ui';
 import { endChat } from '../serverless-calls/endChat';
 
 type Props = {
-  token: string;
   channelSid: string;
+  token: string;
 };
 
 export default function EndChat({ channelSid, token }: Props) {
@@ -13,23 +13,16 @@ export default function EndChat({ channelSid, token }: Props) {
   const handleEndChat = async () => {
     try {
       await endChat(channelSid, token);
-      console.log('await endChat');
     } catch (error) {
       console.log(error);
     }
   };
 
-  // This is a webchat api call to clear the history and open a new window
   const handleExit = async () => {
     await handleEndChat();
-    try {
-      FlexWebChat.Actions.invokeAction('RestartEngagement');
-      window.close();
-
-      window.open('https://google.com');
-    } catch (error) {
-      console.log(error);
-    }
+    // Clear chat history and open a new location
+    await FlexWebChat.Actions.invokeAction('RestartEngagement');
+    setTimeout(() => window.open('https://google.com', '_self'), 1000);
   };
 
   return (
